+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 Changes to Calva.
 
 ## [Unreleased]
+- Fix: [Jack-in is broken for multi-project workspaces](https://github.com/BetterThanTomorrow/calva/issues/821)
 
 ## [2.0.128] - 2020-10-17
-- Fix: [Jack-in is broken](https://github.com/BetterThanTomorrow/calva/issues/821)
+- Fix: [Jack-in is broken if live share extension is not installed](https://github.com/BetterThanTomorrow/calva/issues/821)
 
 ## [2.0.127] - 2020-10-17
 - [Live Share Support](https://github.com/BetterThanTomorrow/calva/issues/803)